Migrate to a plugin based topic system.
