+++ tb2k.py
@@ -5,11 +5,11 @@
 import os
 import sys
 import random
-from datetime import date
 
 import requests
-import wikipedia
 from slackclient import SlackClient
+
+import topics.functions as tf
 
 
 def find_id(channel, bot):
@@ -34,63 +34,6 @@ def find_id(channel, bot):
         sys.exit("tb2k: error: couldn't find #{}".format(channel))
 
 
-def fetch_json(url):
-    """Fetch data from a URL and attempt to parse it as JSON."""
-    try:
-        response = requests.get(url)
-    except requests.ConnectionError:
-        sys.exit("tb2k: error: cannot connect to {}".format(url))
-
-    status_code = response.status_code
-    if status_code == 200:
-        try:
-            return response.json()
-        except ValueError:
-            sys.exit("tb2k: error: invalid JSON at {}".format(url))
-    else:
-        sys.exit("tb2k: error: {} fetching {}".format(status_code, url))
-
-
-def man_page():
-    """Randomly select a man page from /usr/share/man/."""
-    sec = random.randrange(1, 9)
-    man_dir = "/usr/share/man/man{}/".format(sec)
-    man_url = "http://man7.org/linux/man-pages/man{}/".format(sec)
-
-    try:
-        page = random.choice(os.listdir(man_dir))
-    except IndexError:
-        sys.exit("tb2k: error: no man pages in section {}".format(sec))
-    except FileNotFoundError:
-        sys.exit("tb2k: error: {} does not exist".format(man_dir))
-    except PermissionError:
-        sys.exit("tb2k: error: {}: permission denied".format(man_dir))
-
-    man_url = "{}{}.{}.html".format(man_url, page.split('.')[0], sec)
-    return "{}({})".format(page.split('.')[0], sec), man_url
-
-
-def wikipedia_programming_language():
-    """Randomly select a programming language from Wikipedia."""
-    title = 'List of programming languages'
-    page = wikipedia.page(title=title)
-    lang = random.choice(page.links)
-    lang_url = "https://en.wikipedia.org/wiki/{}".format(lang.replace(' ','_'))
-    return lang, lang_url
-
-
-def top_hacker_news_story():
-    """Select the current top Hacker News story."""
-    hn_api = 'https://hacker-news.firebaseio.com/v0'
-
-    stories_url = "{}/topstories.json".format(hn_api)
-    stories = fetch_json(stories_url)
-
-    story_url = "{}/item/{}.json".format(hn_api, stories[0])
-    story = fetch_json(story_url)
-
-    return story['title'], story['url']
-
 def main():
     """
     Set a channel topic to one of:
@@ -107,16 +50,12 @@ def main():
 
     bot = SlackClient(token)
     if bot.rtm_connect():
-        day = date.today().weekday()
-        # Mon Tue Wed
-        if day < 3:
-            topic, link = man_page()
-        # Thur Fri
-        elif day < 5:
-            topic, link = wikipedia_programming_language()
-        # Sat Sun
-        else:
-            topic, link = top_hacker_news_story()
+
+        # Get all attributes in the topics.functions module that begin with
+        # 'topic', randomly choose one, and then execute it.
+        topic_funcs = [getattr(tf, f) for f in dir(tf) if f.startswith('topic')]
+        topic_func = random.choice(topic_funcs)
+        topic, link = topic_func()
 
         channel_id, channel_type = find_id(channel, bot)
