+++ topics/utils.py
@@ -1,0 +1,19 @@
+import sys
+import requests
+
+
+def fetch_json(url):
+    """Fetch data from a URL and attempt to parse it as JSON."""
+    try:
+        response = requests.get(url)
+    except requests.ConnectionError:
+        sys.exit("tb2k: error: cannot connect to {}".format(url))
+
+    status_code = response.status_code
+    if status_code == 200:
+        try:
+            return response.json()
+        except ValueError:
+            sys.exit("tb2k: error: invalid JSON at {}".format(url))
+    else:
+        sys.exit("tb2k: error: {} fetching {}".format(status_code, url))
